Update existing type definitions with restructured imports
