+++ src/types/ContentTypes.ts
@@ -1,4 +1,4 @@
-import { YText, YXmlFragment } from '~/utils/notebookHelpers';
+import { YText, YXmlFragment } from '~/types/YjsTypes';
 
 export type CodeCellContent = YText;
 export type MarkdownCellContent = YXmlFragment;
