+++ src/types/NotebookTypes.ts
@@ -1,5 +1,6 @@
-import { Dispatch, SetStateAction } from 'react';
-import { YMap } from '~/utils/notebookHelpers';
+import { CellType } from "@/Cell";
+import { Dispatch, SetStateAction } from "react";
+import { YMap } from "./YjsTypes";
 
 export type NotebookMetadataType = YMap;
 // & {
@@ -9,7 +10,7 @@ export type NotebookMetadataType = YMap;
 
 export interface NotebookContextType {
   addCellAtIndex: (idx: number, type: string) => void;
-  repositionCell: (cell: any, newIndex: number) => void;
+  repositionCell: (cell: CellType, newIndex: number) => void;
   deleteCell: (id: string) => void;
   allRunning: boolean;
   setAllRunning: Dispatch<SetStateAction<boolean>>;
@@ -19,4 +20,4 @@ export interface NotebookType {
   docID: string;
   title?: string;
   language?: string;
-} 
+}
